requirements.txt
loguru==0.7.2
requests~=2.32.3
nicegui==2.13.0
ruff==0.3.5
langgraph==0.3.11
pydantic~=2.8.2
python-dotenv==1.0.1
langchain==0.3.6
langchain-core~=0.3.12
langchain_community==0.3.4
grpcio==1.71.0
google-genai==1.2.0
google-cloud-storage==3.0.0
setuptools~=75.8.0
protobuf~=5.29.4
tqdm~=4.67.1
tenacity~=9.0.0
pymongo==4.11.2
fastapi==0.115.9
uvicorn==0.34.0
httpx~=0.28.1
langchain-mongodb==0.5.0
langchain-google-genai~=2.0.9
markdown-it-py[plugins]~=3.0.0
beautifulsoup4==4.13.3
slowapi~=0.1.9
redis==5.2.1
chromadb~=1.0.7
matplotlib~=3.9.4
sentence_transformers==4.1.0
langchain-chroma==0.2.3
scikit-learn~=1.6.1